move process instructions to syn crate
