--- lang/src/initialize.rs
@@ -1,69 +1,0 @@
-use solana_program::account_info::{next_account_info, AccountInfo};
-use solana_program::entrypoint::ProgramResult;
-use solana_program::program::invoke_signed;
-use solana_program::program_error::ProgramError;
-use solana_program::pubkey::Pubkey;
-use solana_program::system_instruction;
-use spl_tlv_account_resolution;
-use spl_tlv_account_resolution::account::ExtraAccountMeta;
-use spl_tlv_account_resolution::state::ExtraAccountMetaList;
-use spl_token_2022::extension::StateWithExtensions;
-use spl_token_2022::state::Mint;
-use spl_transfer_hook_interface::error::TransferHookError;
-use spl_transfer_hook_interface::instruction::ExecuteInstruction;
-use spl_transfer_hook_interface::{
-    collect_extra_account_metas_signer_seeds, get_extra_account_metas_address_and_bump_seed,
-};
-
-pub fn process_initialize_extra_account_meta_list(
-    program_id: &Pubkey,
-    accounts: &[AccountInfo],
-    extra_account_metas: &[ExtraAccountMeta],
-) -> ProgramResult {
-    let account_info_iter = &mut accounts.iter();
-
-    let extra_account_metas_info = next_account_info(account_info_iter)?;
-    let mint_info = next_account_info(account_info_iter)?;
-    let authority_info = next_account_info(account_info_iter)?;
-    let _system_program_info = next_account_info(account_info_iter)?;
-
-    let mint_data = mint_info.try_borrow_data()?;
-    let mint = StateWithExtensions::<Mint>::unpack(&mint_data)?;
-    let mint_authority = mint
-        .base
-        .mint_authority
-        .ok_or(TransferHookError::MintHasNoMintAuthority)?;
-
-    if !authority_info.is_signer {
-        return Err(ProgramError::MissingRequiredSignature);
-    }
-    if *authority_info.key != mint_authority {
-        return Err(TransferHookError::IncorrectMintAuthority.into());
-    }
-
-    let (expected_validation_address, bump_seed) =
-        get_extra_account_metas_address_and_bump_seed(mint_info.key, program_id);
-    if expected_validation_address != *extra_account_metas_info.key {
-        return Err(ProgramError::InvalidSeeds);
-    }
-
-    let bump_seed = [bump_seed];
-    let signer_seeds = collect_extra_account_metas_signer_seeds(mint_info.key, &bump_seed);
-    let length = extra_account_metas.len();
-    let account_size = ExtraAccountMetaList::size_of(length)?;
-    invoke_signed(
-        &system_instruction::allocate(extra_account_metas_info.key, account_size as u64),
-        &[extra_account_metas_info.clone()],
-        &[&signer_seeds],
-    )?;
-    invoke_signed(
-        &system_instruction::assign(extra_account_metas_info.key, program_id),
-        &[extra_account_metas_info.clone()],
-        &[&signer_seeds],
-    )?;
-
-    let mut data = extra_account_metas_info.try_borrow_mut_data()?;
-    ExtraAccountMetaList::init::<ExecuteInstruction>(&mut data, extra_account_metas)?;
-
-    Ok(())
-}
