+++ lang/syn/src/codegen/mod.rs
@@ -1,2 +1,5 @@
 pub mod entry;
+pub mod execute;
+pub mod initialize;
 pub mod processor;
+pub mod update;
