+++ lang/syn/src/codegen/processor.rs
@@ -14,12 +14,12 @@ pub fn generate(_program: &TransferHookInput) -> proc_macro2::TokenStream {
             match instruction {
                 TransferHookInstruction::Execute { amount } => {
                     msg!("Instruction: Execute");
-                    kaptn_lang::execute::process_execute(program_id, accounts, amount, process_transfer)
+                    process_execute(program_id, accounts, amount, process_transfer)
                 }
                 TransferHookInstruction::InitializeExtraAccountMetaList { extra_account_metas: _ } => {
                     msg!("Instruction: InitializeExtraAccountMetaList");
                     let user_extra_metas = E::to_extra_account_metas();
-                    kaptn_lang::initialize::process_initialize_extra_account_meta_list(
+                    process_initialize_extra_account_meta_list(
                         program_id,
                         accounts,
                         &user_extra_metas,
@@ -28,7 +28,7 @@ pub fn generate(_program: &TransferHookInput) -> proc_macro2::TokenStream {
                 TransferHookInstruction::UpdateExtraAccountMetaList { extra_account_metas: _ } => {
                     msg!("Instruction: UpdateExtraAccountMetaList");
                     let user_extra_metas = E::to_extra_account_metas();
-                    kaptn_lang::update::process_update_extra_account_meta_list(program_id, accounts, &user_extra_metas)
+                    process_update_extra_account_meta_list(program_id, accounts, &user_extra_metas)
                 }
             }
         }
